+++ FitCloudKit/FitCloudKit.xcframework/ios-arm64_x86_64-simulator/FitCloudKit.framework/Headers/FitCloudDataModels.h
@@ -29,6 +29,8 @@
 #import <FitCloudKit/FitCloudEarbudBatteryInfoModel.h>
 #import <FitCloudKit/FitCloudEarbudsFindStatusInfoModel.h>
 #import <FitCloudKit/FitCloudEarbudsStatusInfoModel.h>
+#import <FitCloudKit/FitCloudWorkoutSlotEditModel.h>
+#import <FitCloudKit/FitCloudHourlyWeatherModel.h>
 
 /// Specific client defines
 #import <FitCloudKit/Huashengda.h>